requirements.txt
black       ==      22.10.0
click     ==        8.1.3
Flask     ==        2.2.2
Flask-Cors   ==     3.0.10
Flask-SQLAlchemy == 3.0.2
greenlet    ==      2.0.1
itsdangerous  ==    2.1.2
Jinja2     ==       3.1.2
MarkupSafe   ==     2.1.1
mypy-extensions  == 0.4.3
mysqlclient  ==     2.1.1
pathspec      ==    0.10.2
pip         ==      22.3.1
platformdirs  ==    2.5.4
python-decouple ==  3.6
python-dotenv  ==   0.21.0
setuptools    ==    65.5.1
six       ==        1.16.0
SQLAlchemy   ==     1.4.44
tomli      ==       2.0.1
Werkzeug    ==      2.2.2
wheel      ==       0.37.1